# Resolve type stability of evaluation of `KernelSum`