--- a/src/kernels/kernelsum.jl
+++ b/src/kernels/kernelsum.jl
@@ -43,22 +43,25 @@ end
 
 Base.length(k::KernelSum) = length(k.kernels)
 
-(κ::KernelSum)(x, y) = sum(k(x, y) for k in κ.kernels)
+_sum(f::Tf, x::Tuple) where {Tf} = f(x[1]) + _sum(f, Base.tail(x))
+_sum(f::Tf, x::Tuple{Tx}) where {Tf, Tx} = f(x[1])
+
+(κ::KernelSum)(x, y) = _sum(k -> k(x, y), κ.kernels)
 
 function kernelmatrix(κ::KernelSum, x::AbstractVector)
-    return sum(kernelmatrix(k, x) for k in κ.kernels)
+    return _sum(Base.Fix2(kernelmatrix, x), κ.kernels)
 end
 
 function kernelmatrix(κ::KernelSum, x::AbstractVector, y::AbstractVector)
-    return sum(kernelmatrix(k, x, y) for k in κ.kernels)
+    return _sum(k -> kernelmatrix(k, x, y), κ.kernels)
 end
 
 function kernelmatrix_diag(κ::KernelSum, x::AbstractVector)
-    return sum(kernelmatrix_diag(k, x) for k in κ.kernels)
+    return _sum(Base.Fix2(kernelmatrix_diag, x), κ.kernels)
 end
 
 function kernelmatrix_diag(κ::KernelSum, x::AbstractVector, y::AbstractVector)
-    return sum(kernelmatrix_diag(k, x, y) for k in κ.kernels)
+    return _sum(k -> kernelmatrix_diag(k, x, y), κ.kernels)
 end
 
 function Base.show(io::IO, κ::KernelSum)