**Summary**
This is an attempt to patch #458. Since this is probably some deeper Julia type inference issue, this solution is likely to be temporary.

**Proposed changes**
For now, I just added tests that expose the type instability. For those kernels, and only or those, the issue can be fixed by including an evaluation of `sum` in the module initialization function.

**What alternatives have you considered?**
Re-writing
https://github.com/JuliaGaussianProcesses/KernelFunctions.jl/blob/ce7923f8252d183768f247256079a2c8279f7b1b/src/kernels/kernelsum.jl#L46
either by using `do` or `let` blocks, or passing an `init` argument to `sum`.
None of those made the tests pass, but they work if the function is redefined in the REPL, after the module has been loaded.

**Breaking changes**
None.