@simsurace 2022-08-30
Hmm, that'd then be similar to the hack I used in https://github.com/JuliaGaussianProcesses/GPLikelihoods.jl/pull/90, right?