@simsurace 2022-08-30
I'd rather not use that here because the allocations might increase too much. A loopy version is probably also not so great for AD.