@simsurace 2022-08-30
Anyway, this seems to work on nightly, presumably due to https://github.com/JuliaLang/julia/pull/45789, but unfortunately that wasn't backported to 1.8.