@simsurace 2022-08-30
Hmm I just did another minimal example on nightly and it didn't pass the test. I guess things might be more complicated after all.
I don't really know how big of a performance issue this is. But hard-coding my kernel in terms of a custom type, where the sum is also hard-coded gave a moderate improvement and made the loss function fully type-stable.