@willtebbutt 2022-09-26
@simsurace have merged in changes from master and re-bumped the patch. Will merge when CI passes.